[trading-composition][Kotlin] Rename the exposed module to db in preparation for the second implementation
