+++ kotlin/settings.gradle.kts
@@ -1,4 +1,4 @@
 rootProject.name = "contract-test-examples"
 
 include("trading:domain", "trading:adapters:exposed", "trading:adapters:in-memory")
-include("trading-composition:domain", "trading-composition:adapters:exposed")
+include("trading-composition:domain", "trading-composition:adapters:db")
